Moved test form to own submodule
