+++ tests/field_tests/views.py
@@ -1,28 +1,11 @@
 """Views of test project."""
 from django.views import View
 from django.views.generic.edit import FormView
-from django.forms import ModelForm, Textarea
+
 from django.http import HttpResponseNotFound, HttpResponseRedirect
 
 from field_tests.models import ExampleTable
-
-
-class ExampleForm(ModelForm):
-    """Form for example table's GVarField."""
-
-    class Meta:
-        """Specification of form model and widget."""
-
-        model = ExampleTable
-        fields = ["a"]
-        widgets = {
-            "a": Textarea(
-                attrs={
-                    "placeholder": "'1(2), 3(4), ...' or '[1, 2] | [[3, 4], [4, 5]]'",
-                    "class": "form-control",
-                },
-            )
-        }
+from field_tests.forms import ExampleForm
 
 
 class IndexView(FormView):
